Django 1.11 cronjob changes (#3532)
* Remove django-cronjobs, replacing with proper management commands

Fixes #3515

* Fix some strptime imports

* Fix some call_command references

* Fix the output of a management command to be text

* Fix the date argument of two new management commands

* Lint fixes

* Fix test syntax error
