+++ kitsune/wiki/management/commands/rebuild_kb.py
@@ -1,0 +1,14 @@
+import waffle
+from django.core.management.base import BaseCommand
+
+from kitsune.wiki import tasks
+
+
+class Command(BaseCommand):
+
+    def handle(self, **options):
+        # If rebuild on demand switch is on, do nothing.
+        if waffle.switch_is_active('wiki-rebuild-on-demand'):
+            return
+
+        tasks.rebuild_kb()
